refactor: internal 패키지로 이동
