+++ member/src/main/java/kr/co/readingtown/member/client/FollowClient.java
@@ -1,6 +1,6 @@
 package kr.co.readingtown.member.client;
 
-import kr.co.readingtown.member.dto.request.FollowBulkCheckRequestDto;
+import kr.co.readingtown.member.dto.request.internal.FollowBulkCheckRequestDto;
 import org.springframework.cloud.openfeign.FeignClient;
 import org.springframework.web.bind.annotation.*;
