+++ member/src/main/java/kr/co/readingtown/member/dto/request/internal/FollowBulkCheckRequestDto.java
@@ -1,4 +1,4 @@
-package kr.co.readingtown.member.dto.request;
+package kr.co.readingtown.member.dto.request.internal;
 
 import jakarta.validation.constraints.NotEmpty;
 import jakarta.validation.constraints.NotNull;
